Rename the ExploredLocation to MineralSite
Changes:
- Renamed explroed locaiton to mineral site to better reflect purpose
- Start moving hte Mission only Task to a new dedicated package

Ref #1599
Ref #1550
