+++ mars-sim-core/src/main/java/com/mars_sim/core/vehicle/task/UnloadHelper.java
@@ -15,8 +15,8 @@ import com.mars_sim.core.building.BuildingManager;
 import com.mars_sim.core.equipment.EVASuit;
 import com.mars_sim.core.equipment.Equipment;
 import com.mars_sim.core.equipment.ResourceHolder;
+import com.mars_sim.core.mission.task.CollectMinedMinerals;
 import com.mars_sim.core.person.Person;
-import com.mars_sim.core.person.ai.task.CollectMinedMinerals;
 import com.mars_sim.core.person.ai.task.CollectResources;
 import com.mars_sim.core.resource.ItemResourceUtil;
 import com.mars_sim.core.resource.Part;
